add     react-native-safe-area-context to package.json. Fix: routing to/from protected routes. Root layout checks for token when mounting
