+++ app/(protected)/index.tsx
@@ -1,9 +1,11 @@
-import { View, Text } from "react-native";
+import { Text } from "react-native";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { styles } from "@/styles/protectedStyles";
 
 export default function Home() {
   return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+    <SafeAreaView style={styles.container}>
       <Text>Protected home</Text>
-    </View>
+    </SafeAreaView>
   );
 }
