+++ app/register/index.tsx
@@ -1,7 +1,7 @@
 // Register
 import { useState } from "react";
 import { Text, TextInput, TouchableOpacity } from "react-native";
-import { SafeAreaView } from "react-native";
+import { SafeAreaView } from "react-native-safe-area-context";
 import { useRouter } from "expo-router";
 import { styles } from "../../styles/registerStyle";
