+++ styles/protectedStyles.ts
@@ -1,0 +1,23 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  button: {
+    color: "white",
+    backgroundColor: "#1DA1F2",
+    padding: 8,
+    borderRadius: 12,
+    width: 250,
+    textAlign: "center",
+    margin: 5,
+  },
+  buttonText: {
+    textAlign: "center",
+    color: "white",
+    fontSize: 16,
+  },
+});
